+++ src/taskprovider.ts
@@ -62,17 +62,9 @@ export function registerTaskProvider( context: vscode.ExtensionContext ) {
 		vscode.tasks.registerTaskProvider( 'bmx', new BmxBuildTaskProvider )
 	)
 
-	// Related commands
+	// Related commands	
 	context.subscriptions.push( vscode.commands.registerCommand( 'blitzmax.build', () => {
 		vscode.tasks.executeTask( makeTask( getBuildDefinitionFromWorkspace( undefined ) ) )
-	} ) )
-
-	context.subscriptions.push( vscode.commands.registerCommand( 'blitzmax.setSourceFile', ( document: vscode.Uri ) => {
-		const path = document ? document.fsPath : vscode.window.activeTextEditor?.document.uri.fsPath
-		if ( !path ) return
-		const def = getBuildDefinitionFromWorkspace()
-		def.source = path
-		saveAsDefaultTaskDefinition( def )
 	} ) )
 }